requirements.txt
Flask
Flask-SQLAlchemy
Flask-WTF
Flask-CKEditor
python-dotenv
gunicorn
